car_bridge/requirements.txt
opencv-python
paho-mqtt
requests
python-dotenv
